Handle cases where negative estimates were found
For some muscles, negative values were found for some parameters, i.e.,
when tendon length does not change in the sampled poses of the reference model.
In these cases, we assume that the same fiber/tendon proportion is found
in the target model as in the reference.

Original changes to Matlab code by Luca Modenese